+++ backend/db/migrate/20220303160442_stage_title_desc.rb
@@ -2,6 +2,6 @@ class StageTitleDesc < ActiveRecord::Migration[6.1]
   def change
     add_column :stages, :title, :string
     add_column :stages, :description, :string
-    add_column :stages, :available_after_days, :integer, null: false, default: 0
+    add_column :stages, :available_after_days, :float, null: false, default: 0
   end
 end
